% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_pca.R
\name{CellCycleScoringPerBatch}
\alias{CellCycleScoringPerBatch}
\title{Score cell cycle phases per batch}
\usage{
CellCycleScoringPerBatch(
  object,
  batch.var = NULL,
  s.features,
  g2m.features,
  ctrl = NULL,
  assay = NULL,
  layer = NULL,
  ...
)
}
\arguments{
\item{object}{A Seurat object}

\item{batch.var}{The name of the batch variable (must be in the object metadata)}

\item{s.features}{A vector of features associated with S phase}

\item{g2m.features}{A vector of features associated with G2M phase}

\item{ctrl}{Number of control features selected from the same bin per
analyzed feature supplied to \code{\link[Seurat]{AddModuleScore}}.
Defaults to value equivalent to minimum number of features
present in 's.features' and 'g2m.features'.}

\item{assay}{assay to use. Passed to Seurat to automatically construct the
\code{batch.var} when not provided. Useless otherwise}

\item{layer}{layer to use. Passed to Seurat to automatically construct the
\code{batch.var} when not provided. Useless otherwise}

\item{...}{Arguments to be passed to \code{\link[Seurat]{CellCycleScoring}},
then \code{\link[Seurat]{AddModuleScore}} (with the exception of
\code{set.ident} which is always \code{FALSE})}
}
\value{
A Seurat object with the following columns added to object meta data: S.Score, G2M.Score, and Phase
}
\description{
Assign cell cycle scores to cells. Scores are computed for each batch
independantly.
}
\examples{
\dontrun{
obj <- SeuratData::LoadData("pbmcsca")
obj[["RNA"]] <- split(obj[["RNA"]], f = obj$Method)
obj <- NormalizeData(obj)
obj <- FindVariableFeatures(obj)
obj <- ScaleData(obj)
obj <- CellCycleScoringPerBatch(obj, batch.var = 'Method',
                                s.features = cc.genes.updated.2019$s.genes,
                                g2m.features = cc.genes.updated.2019$g2m.genes)

head(obj[[]])
}
}
\seealso{
\code{\link[Seurat]{CellCycleScoring}} to compute cc scores on the whole
dataset.
}
